evilginx-ai/requirements.txt
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.20.0
pyyaml>=6.0
dnspython>=2.0.0
cryptography>=40.0.0
jinja2>=3.0.0
python-multipart>=0.0.5

